norm

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: lpittet <lpittet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/02 14:05:32 by lpittet           #+#    #+#             */
-/*   Updated: 2024/10/02 14:24:33 by lpittet          ###   ########.fr       */
+/*   Updated: 2024/10/07 10:57:57 by lpittet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int				i;
+	unsigned long	i;
 	unsigned char	*cs;
 
 	cs = (unsigned char *)s;
