+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: lpittet <lpittet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/02 13:33:47 by lpittet           #+#    #+#             */
-/*   Updated: 2024/10/02 13:39:43 by lpittet          ###   ########.fr       */
+/*   Updated: 2024/10/07 10:58:47 by lpittet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int		i;
-	char	*cdest;
-	char	*csrc;
+	unsigned long	i;
+	char			*cdest;
+	char			*csrc;
 
 	i = 0;
 	cdest = (char *)dest;
